feat: added option to display interactive checkboxes in tasks table
