+++ docs/data-sources/vault-data.md
@@ -43,6 +43,8 @@ Tasks table consists of the following columns:
 | `completed` | 0 if not completed, 1 if completed                                                                                       |               |
 | `path`      | Full path of the file the tag belongs to                                                                                 | 0.24.1        |
 | `filePath`  | (deprecated) same like `path`. Name changed for compatibility with other tables. Will get removed in the future versions |               |
+| `checkbox`  | Interactive checkbox for the task that can be clicked to toggle completion state                                         | 0.29.0        |
+| `position`  | Line number where the task appears in the original file                                                                  | 0.29.0        |
 
 ### `links` table
 Table containing all the links between files.
